requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
